+++ semcore/popper/src/Popper.jsx
@@ -44,6 +44,13 @@ const useUpdatePopperEveryFrame = (popperRef) => {
   return handleAnimationFrame;
 };
 
+const hasParent = (element, parent) => {
+  if (!element) return false;
+  if (element === document.body) return parent === document.body;
+  if (element === parent) return true;
+  return hasParent(element.parentElement, parent);
+};
+
 const MODIFIERS_OPTIONS = [
   'offset',
   'preventOverflow',
@@ -69,6 +76,7 @@ class Popper extends Component {
     interaction: 'click',
     timeout: 0,
     excludeRefs: [],
+    focusLoop: true,
   };
 
   static enhance = [
@@ -360,8 +368,16 @@ class Popper extends Component {
       animationsDisabled,
       popper: this.popper,
       disableEnforceFocus,
+      handleFocusOut: this.handlePopperFocusOut,
     };
   }
+
+  handlePopperFocusOut = (event) => {
+    if (this.asProps.focusLoop) return;
+    if (hasParent(event.relatedTarget, this.triggerRef.current)) return;
+
+    this.bindHandlerChangeVisibleWithTimer(false, 'popper', 'onBlur')(event);
+  };
 
   setContext() {
     return {
@@ -467,13 +483,21 @@ function PopperPopper(props) {
     animationsDisabled,
     popper,
     focusMaster = false,
+    handleFocusOut,
   } = props;
   const ref = React.useRef(null);
 
   // https://github.com/facebook/react/issues/11387
   const stopPropagation = React.useCallback((event) => event.stopPropagation(), []);
 
-  useFocusLock(ref, autoFocus, triggerRef, !visible || disableEnforceFocus, focusMaster);
+  useFocusLock(
+    ref,
+    autoFocus,
+    triggerRef,
+    !visible || disableEnforceFocus,
+    focusMaster,
+    handleFocusOut,
+  );
 
   useContextTheme(ref, visible);